+++ epi/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+* You can control the initial size of the native window with `App::initial_window_size`.
 * You can control the maximum egui web canvas size with `App::max_size_points`.